agetty: move utility functions to utils.c
Move agetty_update_utmp(), agetty_xgethostname() and
agetty_xgetdomainname() to utils.c.

Move the SYSV_STYLE detection to agetty.h so it is available to
all split-out files. Change agetty_update_utmp() to take fakehost
as a parameter instead of accessing a file-scope static variable.
